fix: Moved SupplierInvoice fixture to conftest.py
